# sishotel/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SisHotel",
    "description": "API REST para gestão de reservas de hotel (check-in/check-out). Arquitetura em 3 camadas: Controller → Service → Repository. FIAP - 3ESPR - CP2.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/guests": {
      "get": {
        "tags": [
          "Hóspedes"
        ],
        "summary": "Lista todos os hóspedes",
        "operationId": "list_guests_api_v1_guests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GuestResponse"
                  },
                  "type": "array",
                  "title": "Response List Guests Api V1 Guests Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hóspedes"
        ],
        "summary": "Cadastra novo hóspede",
        "operationId": "create_guest_api_v1_guests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guests/{guest_id}": {
      "get": {
        "tags": [
          "Hóspedes"
        ],
        "summary": "Busca hóspede por ID",
        "operationId": "get_guest_api_v1_guests__guest_id__get",
        "parameters": [
          {
            "name": "guest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hóspedes"
        ],
        "summary": "Atualiza dados do hóspede",
        "operationId": "update_guest_api_v1_guests__guest_id__put",
        "parameters": [
          {
            "name": "guest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guest Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hóspedes"
        ],
        "summary": "Remove um hóspede",
        "operationId": "delete_guest_api_v1_guests__guest_id__delete",
        "parameters": [
          {
            "name": "guest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guest Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reservations": {
      "get": {
        "tags": [
          "Reservas"
        ],
        "summary": "Lista todas as reservas",
        "operationId": "list_reservations_api_v1_reservations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReservationResponse"
                  },
                  "type": "array",
                  "title": "Response List Reservations Api V1 Reservations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reservas"
        ],
        "summary": "Cria nova reserva",
        "operationId": "create_reservation_api_v1_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datas/capacidade inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Hóspede ou quarto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Quarto indisponível no período",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reservations/{reservation_id}": {
      "get": {
        "tags": [
          "Reservas"
        ],
        "summary": "Busca reserva por ID",
        "operationId": "get_reservation_api_v1_reservations__reservation_id__get",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reservations/{reservation_id}/check-in": {
      "post": {
        "tags": [
          "Reservas"
        ],
        "summary": "Realiza check-in da reserva",
        "operationId": "check_in_api_v1_reservations__reservation_id__check_in_post",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Transição de estado inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Fora da janela de check-in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reservations/{reservation_id}/check-out": {
      "post": {
        "tags": [
          "Reservas"
        ],
        "summary": "Realiza check-out da reserva (calcula valor final)",
        "operationId": "check_out_api_v1_reservations__reservation_id__check_out_post",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Transição de estado inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reservations/{reservation_id}/cancel": {
      "post": {
        "tags": [
          "Reservas"
        ],
        "summary": "Cancela reserva (apenas se estiver em CREATED)",
        "operationId": "cancel_api_v1_reservations__reservation_id__cancel_post",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Reserva não está em CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Código identificador do erro"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem explicativa"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Detalhes opcionais (ex.: campos inválidos)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "GuestCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Full Name",
            "description": "Nome completo"
          },
          "document": {
            "type": "string",
            "maxLength": 30,
            "minLength": 5,
            "title": "Document",
            "description": "CPF ou passaporte"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "E-mail válido"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "Telefone com DDD"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "document",
          "email"
        ],
        "title": "GuestCreate"
      },
      "GuestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "document": {
            "type": "string",
            "title": "Document"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "document",
          "email",
          "phone",
          "created_at"
        ],
        "title": "GuestResponse"
      },
      "GuestUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "title": "GuestUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReservationCreate": {
        "properties": {
          "guest_id": {
            "type": "string",
            "title": "Guest Id",
            "description": "UUID do hóspede"
          },
          "room_id": {
            "type": "string",
            "title": "Room Id",
            "description": "UUID do quarto"
          },
          "checkin_expected": {
            "type": "string",
            "format": "date",
            "title": "Checkin Expected",
            "description": "Data prevista de check-in"
          },
          "checkout_expected": {
            "type": "string",
            "format": "date",
            "title": "Checkout Expected",
            "description": "Data prevista de check-out"
          },
          "guests_count": {
            "type": "integer",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Guests Count",
            "description": "Quantidade de hóspedes",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "guest_id",
          "room_id",
          "checkin_expected",
          "checkout_expected"
        ],
        "title": "ReservationCreate"
      },
      "ReservationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "guest_id": {
            "type": "string",
            "title": "Guest Id"
          },
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "checkin_expected": {
            "type": "string",
            "format": "date",
            "title": "Checkin Expected"
          },
          "checkout_expected": {
            "type": "string",
            "format": "date",
            "title": "Checkout Expected"
          },
          "checkin_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin At"
          },
          "checkout_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "estimated_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Amount"
          },
          "final_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Amount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "guest_id",
          "room_id",
          "checkin_expected",
          "checkout_expected",
          "checkin_at",
          "checkout_at",
          "status",
          "estimated_amount",
          "final_amount",
          "created_at",
          "updated_at"
        ],
        "title": "ReservationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}